skillbridge-ai-service/requirements.txt
fastapi
uvicorn
pika
python-dotenv
langchain
langchain-community
langchain-huggingface
sentence-transformers
psycopg2-binary
